+++ include/ParseStruct/Operation.hh
@@ -4,7 +4,7 @@
  * @file Operation.hh
  * @date 2017.08.01
  * @author kmc7468, dev_kr
- * @brief Dlink 코드 파서의 결과가 생성하는 AST 노드들 중 연산과 관련된 노드들을 정의합니다.
+ * @brief Dlink 코드 파서의 결과가 생성하는 추상 구문 트리의 노드들 중 연산과 관련된 노드들을 정의합니다.
  */
 
 #include <cstdint>
@@ -19,81 +19,56 @@
 namespace Dlink
 {
 	/**
-	 * @brief int32_t 타입의 정수를 담는 파싱 노드입니다.
+	 * @brief 32비트 부호 있는 정수 상수를 저장하는 추상 구문 트리의 노드입니다.
+	 * @details 이 구조체는 다른 곳에서 상속받을 수 없습니다.
 	 */
 	struct Integer32 final : public Expression
 	{
-		/**
-		 * @brief int32_t 타입의 정수 데이터입니다.
-		 */
-		std::int32_t data;
-
-		/**
-		 * @brief int32_t를 받아 멤버 필드 data를 초기화하고 Integer32를 생성합니다.
-		 */
-		Integer32(std::int32_t data_) : data(data_)
-		{}
+		Integer32(std::int32_t data_) noexcept;
 
 		std::string tree_gen(std::size_t depth) override;
 		LLVM::Value code_gen() override;
+
+		/** 32비트 부호 있는 정수 상수입니다. */
+		std::int32_t data;
 	};
 }
 
 namespace Dlink
 {
-
 	/**
-	 * @brief 이항 연산의 구조를 담는 파싱 노드입니다.
+	 * @brief 이항 연산의 구조를 담는 추상 구문 트리의 노드입니다.
+	 * @details 이 구조체는 다른 곳에서 상속받을 수 없습니다.
 	 */
 	struct BinaryOperation final : public Expression
 	{
-		/**
-		 * @brief 이항 연산의 연산자를 담는 Operator 타입의 멤버 필드입니다.
-		 */
-		TokenType op;
-
-		/**
-		 * @brief 이항 연산의 좌측 피연산자를 담는 ExpressionPtr 타입의 멤버 필드입니다.
-		 */
-		ExpressionPtr lhs;
-		/**
-		 * @brief 이항 연산의 우측 피연산자를 담는 ExpressionPtr 타입의 멤버 필드입니다.
-		 */
-		ExpressionPtr rhs;
-
-		/**
-		 * @brief Operator와 ExpressionPtr 두 인수를 받아 멤버 필드 op, lhs, rhs를 초기화하고 BinaryOperation을 생성합니다.
-		 */
-		BinaryOperation(TokenType op_, ExpressionPtr lhs_, ExpressionPtr rhs_)
-			: op(op_), lhs(lhs_), rhs(rhs_)
-		{}
+		BinaryOperation(TokenType op_, ExpressionPtr lhs_, ExpressionPtr rhs_);
 
 		std::string tree_gen(std::size_t depth) override;
 		LLVM::Value code_gen() override;
+
+		/** 연산자 타입입니다. */
+		TokenType op;
+		/** 이항 연산에서의 좌측 피연산자입니다. */
+		ExpressionPtr lhs;
+		/** 이항 연산에서의 우측 피연산자입니다. */
+		ExpressionPtr rhs;
 	};
 
 	/**
-	 * @brief 단항 연산의 구조를 담는 파싱 노드입니다.
+	 * @brief 단항 연산의 구조를 담는 추상 구문 트리의 노드입니다.
+	 * @details 이 구조체는 다른 곳에서 상속받을 수 없습니다.
 	 */
 	struct UnaryOperation final : public Expression
 	{
-		/*
-		 * @brief 단항 연산의 연산자를 담는 Operator 타입의 멤버 필드입니다.
-		 */
-		TokenType op;
-		/**
-		 * @brief 단항 연산의 피연산자를 담는 ExpressionPtr 타입의 멤버 필드입니다.
-		 */
-		ExpressionPtr rhs;
-
-		/**
-		 * @brief Operator와 ExpressionPtr를 받아 멤버 필드 op, rhs를 초기화하고 UnaryOperation을 생성합니다.
-		 */
-		UnaryOperation(TokenType op_, ExpressionPtr rhs_)
-			: op(op_), rhs(rhs_)
-		{}
+		UnaryOperation(TokenType op_, ExpressionPtr rhs_);
 
 		std::string tree_gen(std::size_t depth) override;
 		LLVM::Value code_gen() override;
+
+		/** 연산자 타입입니다. */
+		TokenType op;
+		/** 피연산자입니다. */
+		ExpressionPtr rhs;
 	};
 }
