+++ include/ParseStruct/Root.hh
@@ -4,7 +4,7 @@
  * @file Root.hh
  * @date 2017.08.01
  * @author dev_kr, kmc7468
- * @brief Dlink 코드 파서의 결과가 생성하는 AST의 노드들 중 기반이 되는 노드들을 정의합니다.
+ * @brief Dlink 코드 파서의 결과가 생성하는 추상 구문 트리의 노드들 중 기반이 되는 노드들을 정의합니다.
  */
 
 #include <memory>
@@ -18,66 +18,60 @@
 namespace Dlink
 {
 	/**
-	 * @brief 루트 파싱 노드입니다.
+	 * @brief 추상 구문 트리의 루트 노드입니다.
 	 */
 	struct Node
 	{
 		/**
-		 * @brief 현재 노드의 트리형 구조를 std::string 타입으로 시각화합니다.
+		 * @brief 현재 노드의 트리를 std::string 타입으로 시각화합니다.
 		 * @param depth 전체 트리에서 현재 노드의 깊이입니다.
 		 * @return 현재 노드의 트리형 구조를 시각화 시킨 값을 반환합니다.
 		 */
 		virtual std::string tree_gen(std::size_t depth) = 0;
 
 		/**
-		 * @brief 현재 노드의 트리형 구조를 LLVM IR 코드로 만듭니다.
+		 * @brief 현재 노드의 트리를 LLVM IR 코드로 만듭니다.
 		 * @return 파싱 노드에서 생성한 LLVM::Value를 반환합니다. 생성한 값이 없을 경우 nullptr을 반환합니다.
 		 */
 		virtual LLVM::Value code_gen() = 0;
 	};
 
 	/**
-	 * @brief Dlink 코드의 식입니다.
+	 * @brief Dlink 코드에서의 식입니다.
 	 */
 	struct Expression : public Node
 	{};
 
 	/**
-	 * @brief Dlink 코드에서 한개 이상의 Expression으로 이루어진 문입니다.
+	 * @brief Dlink 코드에서의 문입니다.
 	 */
 	struct Statement : public Node
 	{};
 
 	/**
-	 * @brief Dlink의 자료형입니다.
+	 * @brief Dlink의 타입(자료형)입니다.
 	 */
 	struct Type
 	{
 		/**
-		 * @brief 현재 타입 노드의 트리형 구조를 std::string 타입으로 시각화합니다.
+		 * @brief 현재 타입 노드의 트리를 std::string 타입으로 시각화합니다.
 		 * @param depth 전체 트리에서 현재 노드의 깊이입니다.
-		 * @return 현재 노드의 트리형 구조를 시각화 시킨 값을 반환합니다.
+		 * @return 현재 노드의 트리를 시각화 시킨 값을 반환합니다.
 		 */
 		virtual std::string tree_gen(std::size_t depth) = 0;
 
 		/**
-		 * @brief 현재 타입 노드를 LLVM의 Type 객체로 만듭니다.
-		 * @return 현재 타입 노드의 트리형 구조를 시각화 시킨 값을 반환합니다.
+		 * @brief 현재 타입 노드를 LLVM Type으로 만듭니다.
+		 * @return 만들어진 LLVM Type을 반환합니다.
 		 */
 		virtual llvm::Type* get_type() = 0;
 	};
 
-	/**
-	 * @brief Expression 구조체에 대한 std::shared_ptr 타입입니다.
-	 */
+	/** Expression 구조체에 대한 std::shared_ptr 타입입니다. */
 	using ExpressionPtr = std::shared_ptr<Expression>;
-	/**
-	 * @brief Statement 구조체에 대한 std::shared_ptr 타입입니다.
-	 */
+	/** Statement 구조체에 대한 std::shared_ptr 타입입니다. */
 	using StatementPtr = std::shared_ptr<Statement>;
-	/**
-	 * @brief Type 구조처에 대한 std::shared_ptr 타입입니다.
-	 */
+	/** Type 구조처에 대한 std::shared_ptr 타입입니다. */
 	using TypePtr = std::shared_ptr<Type>;
 }
 
@@ -85,88 +79,64 @@ namespace Dlink
 {
 	/**
 	 * @brief 식별자입니다.
+	 * @details 이 구조체는 다른 곳에서 상속받을 수 없습니다.
 	 */
 	struct Identifer final : public Expression
 	{
-		/**
-		 * @brief 실질직인 식별자 값입니다.
-		 */
-		const std::string id;
-	
-		/**
-		 * @brief 새 Identifier 객체를 초기화 합니다.
-		 * @param id_ 실질적인 식별자 값입니다.
-		 */
-		Identifer(const std::string& id_)
-			: id(id_)
-		{}
+		Identifer(const std::string& id_);
 
 		std::string tree_gen(std::size_t depth) override;
 		LLVM::Value code_gen() override;
+
+		/** 실질적인 식별자 값입니다. */
+		const std::string id;
 	};
 
 	/**
-	 * @brief 한 개 이상의 Statement들의 집합입니다.
+	 * @brief 한개 이상의 Statement들의 집합입니다.
 	 */
 	struct Block : public Statement
 	{
-		/**
-		 * @brief Statement들을 담아두는 StatementPtr의 std::vector입니다.
-		 */
-		std::vector<StatementPtr> statements;
-
-		/**
-		 * @brief StatementPtr의 std::vector를 받아 멤버 필드 statements를 초기화하고 Block을 생성합니다.
-		 */
-		Block(std::vector<StatementPtr> statements_) : statements(statements_)
-		{}
+		Block(const std::vector<StatementPtr>& statements);
 
 		std::string tree_gen(std::size_t depth) override;
 		LLVM::Value code_gen() override;
+
+		/** Statemenet들의 집합입니다. */
+		std::vector<StatementPtr> statements;
 	};
-	
+
 	/**
 	 * @brief Dlink 코드에서 중괄호에 의해 감싸지며 객체의 수명을 결정하는 Statement들의 집합입니다.
+	 * @details 이 구조체는 다른 곳에서 상속받을 수 없습니다.
 	 */
 	struct Scope final : public Block
 	{
-		/**
-		 * @brief 현재 Scope의 부모 Block 또는 부모 Scope입니다.
-		 */
-		StatementPtr parent;
-
-		/**
-		 * @brief 현재 Scope의 자식 Scope입니다.
-		 */
-		std::vector<std::shared_ptr<Scope>> child;
-
-		using Block::Block;
+		Scope(const std::vector<StatementPtr>& statements, StatementPtr parent);
 
 		std::string tree_gen(std::size_t depth) override;
 		LLVM::Value code_gen() override;
+
+		/** 현재 Scope의 상위 Block 또는 상위 Scope입니다.  */
+		StatementPtr parent;
+		/** 현재 Scope의 하위 Scope입니다. */
+		std::vector<std::shared_ptr<Scope>> child;
 	};
-	/**
-	 * @brief Scope 구조체에 대한 std::shared_ptr 타입입니다.
-	 */
+	/** Scope 구조체에 대한 std::shared_ptr 타입입니다. */
 	using ScopePtr = std::shared_ptr<Scope>;
 
 	/**
-	 * @brief 한 개의 Expression으로 이루어진 Statement입니다.
+	 * @brief 한 개의 식으로 이루어진 Statement입니다.
+	 * @details 이 구조체는 다른 곳에서 상속받을 수 없습니다.
 	 */
 	struct ExpressionStatement final : public Statement
 	{
-		/**
-		 * @brief Expression을 담아두는 ExpressionPtr 타입의 멤버 필드입니다.
-		 */
-		ExpressionPtr expression;
-
-		/**
-		 * @brief ExpressionPtr을 받아 멤버 필드 expression을 초기화하고 ExpressionStatement를 생성합니다.
-		 */
-		ExpressionStatement(ExpressionPtr expression_) : expression(expression_)
-		{}
+		ExpressionStatement(ExpressionPtr expression_);
 
 		std::string tree_gen(std::size_t depth) override;
 		LLVM::Value code_gen() override;
+
+		/** 식입니다. */
+		ExpressionPtr expression;
 	};
 }
