+++ include/ParseStruct/Type.hh
@@ -4,7 +4,7 @@
  * @file Type.hh
  * @date 2017.08.01
  * @author dev_kr, kmc7468
- * @brief Dlink 코드 파서의 결과가 생성하는 AST 노드들 중 타입과 관련된 노드들을 정의합니다.
+ * @brief Dlink 코드 파서의 결과가 생성하는 추상 구문 트리 노드들 중 타입과 관련된 노드들을 정의합니다.
  */
 
 #include <string>
@@ -17,23 +17,17 @@
 namespace Dlink
 {
 	/**
-	 * @brief int, char와 같이 한정자나 포인터 선언 등이 없는 순수 타입입니다.
+	 * @brief 한정자 등이 없는 순수한 타입입니다.
+	 * @details 이 구조체는 다른 곳에서 상속받을 수 없습니다.
 	 */
 	struct SimpleType final : public Type
 	{
-		/**
-		 * @brief 타입의 문자열 식별자를 담는 Identifer 타입의 멤버 필드입니다.
-		 */
-		const Identifer identifier;
-
-		/**
-		 * @brief std::string을 받아 멤버 필드 identifier를 초기화하고 SimpleType을 생성합니다.
-		 */
-		SimpleType(const std::string& identifier_)
-			: identifier(identifier_)
-		{}
+		SimpleType(const std::string& identifier_);
 
 		std::string tree_gen(std::size_t depth) override;
 		llvm::Type* get_type() override;
+
+		/** 타입의 식별자입니다. */
+		const Identifer identifier;
 	};
 }
