+++ src/CodeGen.cc
@@ -9,11 +9,8 @@ namespace Dlink
 {
 	namespace LLVM
 	{
-		/**  */
 		llvm::LLVMContext context;
-		/**  */
 		std::shared_ptr<llvm::Module> module = std::make_shared<llvm::Module>("top", context);
-		/**  */
 		llvm::IRBuilder<> builder(context);
 	}
 
@@ -59,128 +56,4 @@ namespace Dlink
 	SymbolTablePtr symbol_table = std::make_shared<SymbolTable>();
 	/** 현재 사용자 정의 타입 심볼 테이블입니다. */
 	TypeSymbolTablePtr type_symbol_table = std::make_shared<TypeSymbolTable>();
-}
-
-namespace Dlink
-{
-	LLVM::Value Identifer::code_gen()
-	{
-		LLVM::Value result = symbol_table->find(id);
-
-		if (result == nullptr)
-		{
-			// TODO
-			return nullptr;
-		}
-
-		return result;
-	}
-
-	LLVM::Value Block::code_gen()
-	{
-		for (StatementPtr statement : statements)
-		{
-			statement->code_gen();
-		}
-
-		return nullptr;
-	}
-
-	LLVM::Value Scope::code_gen()
-	{
-		for (StatementPtr statement : statements)
-		{
-			statement->code_gen();
-		}
-
-		return nullptr;
-	}
-
-	LLVM::Value ExpressionStatement::code_gen()
-	{
-		return expression->code_gen();
-	}
-}
-
-namespace Dlink
-{
-	LLVM::Value Integer32::code_gen()
-	{
-		return LLVM::builder.getInt32(data);
-	}
-
-	LLVM::Value BinaryOperation::code_gen()
-	{
-		LLVM::Value lhs_value = lhs->code_gen();
-		LLVM::Value rhs_value = rhs->code_gen();
-
-		switch (op)
-		{
-		case TokenType::plus:
-			return LLVM::builder.CreateAdd(lhs_value, rhs_value);
-
-		case TokenType::minus:
-			return LLVM::builder.CreateSub(lhs_value, rhs_value);
-
-		case TokenType::multiply:
-			return LLVM::builder.CreateMul(lhs_value, rhs_value);
-
-		case TokenType::divide:
-			// TODO: 임시 방안
-			return LLVM::builder.CreateSDiv(lhs_value, rhs_value);
-
-		default:
-			// TODO: 오류 처리
-			return LLVM::builder.getFalse();
-		}
-	}
-
-	LLVM::Value UnaryOperation::code_gen()
-	{
-		LLVM::Value rhs_value = rhs->code_gen();
-
-		switch (op)
-		{
-		case TokenType::plus:
-			return LLVM::builder.CreateMul(LLVM::builder.getInt32(1), rhs_value);
-
-		case TokenType::minus:
-			return LLVM::builder.CreateMul(LLVM::builder.getInt32(-1), rhs_value);
-
-		default:
-			// TODO: 오류 처리
-			return LLVM::builder.getFalse();
-		}
-	}
-}
-
-namespace Dlink
-{
-	llvm::Type* SimpleType::get_type()
-	{
-		if (identifier.id == "int")
-		{
-			return LLVM::builder.getInt32Ty();
-		}
-
-		return nullptr;
-	}
-}
-
-namespace Dlink
-{
-	LLVM::Value VariableDeclaration::code_gen()
-	{
-		llvm::AllocaInst* var = LLVM::builder.CreateAlloca(type->get_type(), nullptr, identifier.id);
-		var->setAlignment(4);
-
-		if (expression)
-		{
-			LLVM::Value init_expr = expression->code_gen();
-			LLVM::builder.CreateStore(init_expr, var);
-		}
-
-		symbol_table->map.insert(std::make_pair(identifier.id, var));
-		return var;
-	}
 }
